integrating contact creatio and adding UX interations

+++ api/src/app/controllers/ContactController.js
@@ -1,3 +1,4 @@
+const isValidUUID = require('../../utils/isValidUUID');
 const ContactsRepo = require('../repos/ContactsRepo');
 
 class ContactController {
@@ -11,6 +12,10 @@ class ContactController {
 	async show(req, res) {
 
 		const { id } = req.params;
+
+		if (!isValidUUID(id)) {
+			return res.status(400).json({error: "Invalid contact ID"})
+		}
 		const contact = await ContactsRepo.findById(id)
 
 		if (!contact) {
@@ -25,38 +30,57 @@ class ContactController {
 		if (!name) {
 			return res.status(400).json({error: "Name required"})
 		}
-
-		const contactExists = await ContactsRepo.findByEmail(email);
-		if (contactExists) {
-			return res.status(400).json({error: "Contact email already exists"})
+		if (category_id && !isValidUUID(category_id)) {
+			return res.status(400).json({error: "Invalid category"})
 		}
 
-		const contact = await ContactsRepo.create({ name, phone, email, category_id });
+		if (email) {
+			const contactExists = await ContactsRepo.findByEmail(email);
+			if (contactExists) {
+			return res.status(400).json({error: "Contact email already exists"})
+			}
+		}
+
+		const contact = await ContactsRepo.create({
+			name,
+			phone,
+			email: email || null,
+			category_id: category_id || null,
+		});
 		res.status(201).json(contact)
 
 	}
 
 	async update(req, res) {
 		const { id } = req.params;
-		const { name, phone, email, category_id } = req.body;
-		const contactExists = await ContactsRepo.findById(id)
-		if (!contactExists) {
-			return res.status(404).json({error: "User not found"})
+
+		if (!isValidUUID(id)) {
+			return res.status(400).json({ error: "Invalid contact ID" })
 		}
-		const emailExists = await ContactsRepo.findByEmail(email)
-		if (emailExists && emailExists.id !== id) {
-			return res.status(404).json({error: "E-mail already in use"})
+		if (category_id && !isValidUUID(category_id)) {
+			return res.status(400).json({ error: "Invalid category" })
 		}
 
+		const { name, phone, email, category_id } = req.body;
 		if (!name) {
-			return res.status(400).json({error: "Name required"})
+			return res.status(400).json({ error: "Name required" })
+		}
+		const contactExists = await ContactsRepo.findById(id)
+		if (!contactExists) {
+			return res.status(404).json({ error: "User not found" })
+		}
+		if (email) {
+			const emailExists = await ContactsRepo.findByEmail(email)
+			if (emailExists && emailExists.id !== id) {
+				return res.status(404).json({ error: "E-mail already in use" })
+			}
 		}
 
 		const updatedContact = await ContactsRepo.update(id, {
 			name,
-			email,
+			email: email || null,
 			phone,
-			category_id,
+			category_id: category_id || null,
 		});
 		res.json(updatedContact)
 	}
@@ -65,9 +89,9 @@ class ContactController {
 		const { id } = req.params;
 		// const contact = await ContactsRepo.findById(id)
 
-		// if (!contact) {
-		// 	return res.status(404).json({error: "User not found"})
-		// }
+		if (!isValidUUID(id)) {
+			return res.status(400).json({error: "Invalid contact ID"})
+		}
 
 		await ContactsRepo.delete(id);
 		res.sendStatus(204)
